DP-23393: Moving the mass_entity_hierarchy work to the mass_hierarchy module to consolidate efforts with other work
