--- docroot/modules/custom/mass_entity_hierarchy/mass_entity_hierarchy.info.yml
@@ -1,9 +1,0 @@
-name: Mass.gov Entity Hierarchy module
-type: module
-description: Overrides for the Entity Hierarchy module.
-core: 8.x
-core_version_requirement: ^8 || ^9
-package: Custom
-dependencies:
-  - entity_hierarchy
-  - entity_hierarchy_breadcrumb
